Add support for parsing files in the CLI
Create commands in the ctl currently require each field to be specified
via a flag. This can become tedious and error prone IME. To make this
more ergonomic, I've added support for `-f` which accepts either a file
path or the literal `-` which signifies that stdin should be used. Any
values supplied on the CLI will override values read in from the file.